requirements.txt
Django==1.11.6
djangorestframework==3.7.1
django-phonenumber-field==1.3.0
django-filter==1.1.0
django-rest-swagger==2.1.2
pyaml==17.10.0
